+++ crates/gen-rust-wasm/tests/codegen.rs
@@ -28,10 +28,5 @@ mod exports {
         // generator just yet
         "!wasi_next.witx"
         "!host.witx"
-
-        // These use the preview1 ABI which isn't implemented for rust_wasm exports.
-        "!wasi_snapshot_preview1.witx"
-        "!typenames.witx"
-        "!legacy.witx"
     );
 }
