+++ tests/runtime/invalid/host.rs
@@ -1,4 +1,4 @@
-witx_bindgen_wasmtime::export!("./tests/runtime/invalid/imports.witx");
+witx_bindgen_wasmtime::import!("./tests/runtime/invalid/imports.witx");
 
 use anyhow::Result;
 use imports::*;
@@ -36,7 +36,7 @@ impl Imports for MyImports {
     }
 }
 
-witx_bindgen_wasmtime::import!("./tests/runtime/invalid/exports.witx");
+witx_bindgen_wasmtime::export!("./tests/runtime/invalid/exports.witx");
 
 fn run(wasm: &str) -> Result<()> {
     use exports::*;
