+++ tests/runtime/numbers/host.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 
-witx_bindgen_wasmtime::export!("./tests/runtime/numbers/imports.witx");
+witx_bindgen_wasmtime::import!("./tests/runtime/numbers/imports.witx");
 
 #[derive(Default)]
 pub struct MyImports {
@@ -61,7 +61,7 @@ impl imports::Imports for MyImports {
     }
 }
 
-witx_bindgen_wasmtime::import!("./tests/runtime/numbers/exports.witx");
+witx_bindgen_wasmtime::export!("./tests/runtime/numbers/exports.witx");
 
 fn run(wasm: &str) -> Result<()> {
     let (exports, mut store) = crate::instantiate(
